requirements.txt
requests
sqlite-utils
fastapi
uvicorn
python-dotenv
aiohttp
numpy
pydantic
beautifulsoup4
html2text
tqdm
markdown
python-multipart
setuptools
markdownify
playwright
requests
google-generativeai 
pillow
openai